Create internal secret to consume in sub-crs
Create an internal secret in the watcher controller to pass to the
sub CRs (atm only for WatcherAPI). Initially, this secret contains the
service password and the transport url. Eventually, this secret will be
set as the Secret field in the sub CRs specs, once they are created from
the Watcher controller.