Create WatcherAPI keystoneEndpoint
Create the WatcherAPI keystoneEndpoint. After this point, we can run
watcher commands from the openstackclient pod (even though we can't do
anything because the decision engine and applier are not available yet).

This version does not support TLS, so the control plane must be deployed
with TLS disabled as well.