Prometheus integration in Watcher controller
Watcher will use Prometheus as metrics store. We need to integrate the
prometheus configuration from the Watcher controller. I am proposing a
similar approach that the one in the autoscaling controller in the
Telemetry operator.

New parameters have been added to the Watcher spec which allow to
specify a custom Prometheus server:

- PrometheusHost: the fqdn of the prometheus API.
- PrometheusPort: the port number of the prometheus API.
- PrometheusTLSCaCertSecret: a secretkeyselector object pointing to a
  existing secret that needs to be mounted in case that the prometheus
  API has TLS enabled.

In case these values are not set explicitely in the Watcher CR, it will
be assume that it is deployed via the integrated MetricStorage object
and the host, port and TLS config is discovered. In this case, if a
MetricStorage object is not running, the Watcher deployment wil fail.

The Watcher controller will add the prometheus connection parameters to
the SubCrsSecret object so that it can be easily included in the
configuration of 2nd level controllers. In the case of PrometheusTLSCaCertSecret
it is added to the SubCRDs as it is a more convenient way to access it
to mount it as a volume when needed.

Note that this change is also adding TLS option to all the CRDs as it
will be required to connect to the integrated prometheus which enables
TLS by default.