Default watcher container images to master release
Currently watcher-operator defaults to antelope release container
images. However, in order to integrate in openatck-k8s-operators
environments it requires integration with prometheus which will only be
supported in master (epoxy) cycle.

This patch is setting master container images as defaults. Once epoxy is
release we can move to them if available.