Reenable tls in the kuttl test control plane
I disabled tls in the control plane bacause the watcherapi pod could not
talk to keystone since it was lacking the CA to establish an SSL
connection. This was maybe unnecessary, but since now we have the CI
injected in the watcherapi pod, we can use TLS in the controlplane
without any problem.