Use the secret name returned from createSubLevelSecret
The createSubLevelSecret returns the name of the created secret. This
was until now unused, and when passing the secret to the WatcherAPI
spec, we were using the instance name directly, instead of using the
name returned by the method. This change adds the secret name as input
to the ensure API method, so that we ensure we'll use the right secret
name if we ever were to change it in createSubLevelSecret.