Add prometheusSecret into the Spec for prometheus configuration
In order to configure Prometheus datasource in Watcher we are using a
secret which will provide the required connection data. This secret will
be provided by the telemetry operator at some point although could be
any secret which has the expected fields, host, port and
ca_secret and ca_key.

This patch adds the new field into the spec, validates that it provides the
required fields and copy from Watcher to SubCRs. It also adds a watcher
on the secret in the main controller (also adds the watcher on the top
level .Spec.secret which was misssing).