Support TLS in watcher-operator
Duplicate the TLS support code in openstack-operator for TLS support in
watcher-operator before it's integrated. This will support TLS to the
pod level, but only using certificates pre-created.

Check the input corresponds to a valid TLS mode, e.g. when the user
has set the secrets to configure PodLevel TLS, ensure that the secret
to configure Ingress TLS is also set.