Integrate WatcherDecisionEngine into top level Watcher controller
This adds the WatcherDecisionEngine deployment into the main watcher
Reconcile. It introduces the WatcherDecisionEngineTemplate struct
and adds this into the WatcherTemplate. Functional tests now include
validation for the DecisionEngine stateful set and expected defaults.
The DecisionEngine kuttl test scenario is removed and the assertion for
DecisionEngine is moved into the main watcher scenario.

Related: https://issues.redhat.com/browse/OSPRH-11484

	kubeconfig