Reorganize configuration templates
Before this patch all the config files were in the watcher directory and
all the rest were symlinks to it. While this worked fine, that
implementation leads to config-data secrets having unneeded data.

This patch is implementing a similar approach to nova-operator where the
common config file is in the root templates directory while the rest are
in the specific controller folder where they are needed.