remove REXML documemtion dependency
REXML was previously pinned to a version with several
know cves related to regex performance.

we do not directly use ruby or REXML in this project and
as such this is a indirect dependecy of the tool we use
to build docs.

we shoudl not manage asciidoctor's deps for it by pining
REXML, so this change removes it from the GEMFILE