Add information to the Watcher service deployment section
When creating a Watcher CR, certain values of the spec need to match the
values in the existing Control Plane. Although default values in Watcher
are aligned with the values in th openstack operator, they may be
customized.

I'm adding some information about the values that need to match and how
to find the existing ones.