Remove custom nodesets and use inherited one
After [1] merged the parent ci-framework jobs are using 4.18 crc cloud
nodeset, this change removes the custom nodeset and uses the one from
the parent job.

Since the 4.16 job has to use an extracted nodeset and is not working at
the moment of  createing this commit, this change also sets it to
non-voting until it's working again.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/2819