Add isReady method to Watcher and SubCRs
isReady function is consistently implemented in other openstack
operators as a way to easily find if a CR has ReadyCondition true.
It is used, for example, to be called from openstack-operator when
deploying a CR as part of the controlplane.

This patch is adding the new method to Watcher, WatcherAPI,
WatcherApplier and WatcherDecisionEngine. It also adds some functional
tests.

Jira: https://issues.redhat.com/browse/OSPRH-16237