Rename WatcherTemplate to WatcherSpecCore
To be consistent with the convention in openstack operators to be
integrated in the openstack-operator.

Note that, the content of the struct is exactly what we need to import
inside the watcher template definition for the control plane, which is
the entire CRD struct except the images part which come from the
openstackversion object.

Currently, it has some content related to route definition which will be
removed before integrating it into the openstack-operator.

Jira: https://issues.redhat.com/browse/OSPRH-16235