Patch openstackversion when deploying in integrated mode
When deploying in integrated mode we need to patch the openstackversion
object to deploy watcher using the desired watcher container images
which may be built in parent content provider job.

Note that, at some point we will be able to do this directly in
ci-framework using the update_containers role [1], but we may break jobs
running previous versions of openstack-operators doing it already.

[1] https://github.com/openstack-k8s-operators/ci-framework/blob/08aabb7a73daac8fe82922e3551ea1c6f1c5266c/roles/update_containers/templates/update_containers.j2

Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>