Skip volume_migration to avoid failures in check
Patch [1] adds tests for volume migrations using the zone migration
strategy. The test can't configure a wait until the storage model is
updated and thus often fails because it's missing some volume as shown
in [2]. Until we have Cinder notifications enabled, which solves the
problem upstream, skip the test so it does not run and thus unblocks
merging the upstream tempest-plugin patch.

[1] https://review.opendev.org/c/openstack/watcher-tempest-plugin/+/954625
[2] https://github.com/openstack-k8s-operators/watcher-operator/pull/217