[WIP] Enable Ceph as cinder backend with a single pool
Deploy an HCI configuration using cifmw and use it as backend
for cinder. This is needed to test volume migrations and depending on
the result we might enable it in all jobs or only in some.

Deploy with only one pool to test that the test is skipped if there are
not at least two available pools to cinder.

Depends-On: https://review.opendev.org/c/openstack/watcher-tempest-plugin/+/958644
Depends-On: https://review.opendev.org/c/openstack/watcher/+/960265
Depends-On: https://github.com/openstack-k8s-operators/watcher-operator/pull/269